Blocked put and get request for normal user

+++ pages/api/products/[id].js
@@ -4,7 +4,11 @@ export default async function handler(req, res) {
     const {
         method,
         query: { id },
+        cookies,
     } = req;
+
+    const token = cookies.token;
+
     dbConnect();
     if (method === "GET") {
         try {
@@ -15,6 +19,9 @@ export default async function handler(req, res) {
         }
     }
     if (method === "PUT") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("Not authenticated!");
+        }
         try {
             const product = await Product.create(req.body);
             res.status(201).json(product);
@@ -23,6 +30,9 @@ export default async function handler(req, res) {
         }
     }
     if (method === "DELETE") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("Not authenticated!");
+        }
         try {
             await Product.findByIdAndDelete(id);
             res.status(200).json("The product has been deleted!");
